fix: change team_id column type from unsignedBigInteger to integer in users table migration

+++ database/migrations/2026_06_10_000002_add_team_id_to_users_for_qb.php
@@ -10,7 +10,7 @@ return new class extends Migration
     public function up(): void
     {
         Schema::table('users', function (Blueprint $table) {
-            $table->unsignedBigInteger('team_id')->nullable()->after('league_id');
+            $table->integer('team_id')->nullable()->after('league_id');
             $table->foreign('team_id')->references('id')->on('league_teams')->nullOnDelete();
         });
